requirements.txt
fastapi
python-dotenv
pydantic-settings
PyJWT
uvicorn[standard]
supabase